% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agricultural_expansion.R
\name{agricultural_expansion}
\alias{agricultural_expansion}
\title{Perform agricultural expansion scenario}
\usage{
agricultural_expansion(lcm_path)
}
\arguments{
\item{lcm_path}{A character string giving the path to the lcm csv}
}
\description{
Perform agricultural expansion scenario
}
