% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_layer.R
\name{crop_layer}
\alias{crop_layer}
\title{Crop a raster layer given a postcode}
\usage{
crop_layer(postcode = "OX108BB", radius_m = 5000, layer = NULL)
}
\arguments{
\item{postcode}{A character string giving a valid UK postcode}

\item{layer}{A raster layer to be cropped}

\item{radius}{Numeric in meters giving the radius of the bounding box}
}
\value{
A raster layer
}
\description{
Function to crop a raster layer, using the bounding box of a specified radius
(in meters) around a point (centroid of a postcode).
}
