% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_linear_layer.R
\name{crop_linear_layer}
\alias{crop_linear_layer}
\title{Crop a linear layer given a postcode}
\usage{
crop_linear_layer(postcode = "OX108BB", radius_m = 5000, layer = NULL,
  buffer = 15, out_type = 100)
}
\arguments{
\item{postcode}{A character string giving a valid UK postcode}

\item{layer}{A layer to be cropped (ie road or river)}

\item{buffer}{Buffer in meters to add to the lines}

\item{out_type}{A numeric code to give cells containing this line type (e.g. 100 = motorway, 200 = other roads)}

\item{radius}{Numeric in meters giving the radius of the bounding box}
}
\value{
A raster layer
}
\description{
Function to linear layer, using the bounding box of a specified radius
(in meters) around a point (centroid of a postcode).
}
