% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_road_layer.R
\name{crop_road_layer}
\alias{crop_road_layer}
\title{Crop a road layer given a postcode}
\usage{
crop_road_layer(postcode = "OX108BB", radius_m = 5000, roads = NULL,
  buffer = 15, road_type = 100)
}
\arguments{
\item{postcode}{A character string giving a valid UK postcode}

\item{buffer}{Buffer in meters to add to the roads}

\item{road_type}{A numeric code to give cells containing this road type (100 = motorway, 200 = other roads)}

\item{radius}{Numeric in meters giving the radius of the bounding box}

\item{layer}{A road layer to be cropped}
}
\value{
A raster layer
}
\description{
Function to crop a road layer, using the bounding box of a specified radius
(in meters) around a point (centroid of a postcode).
}
