% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_raster.R
\name{format_raster}
\alias{format_raster}
\title{Convert rasters for mapping}
\usage{
format_raster(lcm, dtm, name = "MyWorld", outDir = tempdir(),
  exagerate_elevation = 2)
}
\arguments{
\item{lcm}{Raster of the Land Cover Map}

\item{dtm}{Raster digital terrain model}

\item{name}{Character, the name of the map, ususally the location}

\item{outDir}{Character, the location the save the results, usually a temporary location}

\item{exagerate_elevation}{How much should elevation be exaggerated by?}
}
\value{
A list of length 2. first element is the path to the dtm, second is the lcm
}
\description{
This function does a number of tasks to make the maps ready
for building into a minecraft world. This includes translating
the lcm types.
}
