% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoaddress_uk.R
\name{geoaddress_uk}
\alias{geoaddress_uk}
\title{Get location from a postcode}
\usage{
geoaddress_uk(address, format = ifelse(length(address) == 1, "list",
  "dataframe"))
}
\arguments{
\item{address}{The postcode to search for}

\item{format}{The format postcodes are given in  ('list' or 'data.frame')}
}
\value{
A dataframe of location information
}
\description{
function to geolocalize a postcode or find a centroid of a location
}
