% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview.R
\name{overview}
\alias{overview}
\title{Create and Overview map of a saved minecraft world}
\usage{
overview(savePath, outPath = ".", launch = TRUE)
}
\arguments{
\item{savePath}{the path to the folder containing the saved game to be mapped}

\item{outPath}{the path to a folder where a host of output files will be saved}

\item{launch}{Logical, should the map be launched once it is complete}
}
\value{
Path to the overview index.html
}
\description{
This uses the overviewer programme which is bundled in with the package.
This is currently the windows version so will not work on Linux/Mac
}
