% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seminatural_expansion.R
\name{seminatural_expansion}
\alias{seminatural_expansion}
\title{Perform semi-natural habitat expansion scenario}
\usage{
seminatural_expansion(lcm_path, seminatural_expansion = 0.2,
  improved_grassland_expansion = 0.1, woodland_expansion = 0.05,
  prop_winter_wheat = 0.2, prop_oilseed_rape = 0.1, prop_barley = 0.2,
  prop_field_bean = 0.05, prop_potato = 0.05, prop_sugar_beet = 0.05,
  prop_maize = 0.05)
}
\arguments{
\item{lcm_path}{A character string giving the path to the lcm csv}

\item{seminatural_expansion}{0.2}

\item{improved_grassland_expansion}{0.1}

\item{woodland_expansion}{0.05}

\item{prop_winter_wheat}{0.20}

\item{prop_oilseed_rape}{0.10}

\item{prop_barley}{0.20}

\item{prop_field_bean}{0.05}

\item{prop_potato}{0.05}

\item{prop_sugar_beet}{0.05}

\item{prop_maize}{0.05}
}
\description{
Perform semi-natural habitat expansion scenario
}
