% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcode_map.R
\name{postcode_map}
\alias{postcode_map}
\title{Create Minecraft world from Postcode}
\usage{

  postcode_map(lcm_raster = raster::raster("W:/PYWELL_SHARED/Pywell Projects/BRC/Tom August/Minecraft/base_layers/landcover_composite_map.tif"),
  dtm_raster = raster::raster("W:/PYWELL_SHARED/Pywell Projects/BRC/Tom August/Minecraft/base_layers/uk_elev25.tif"),
  postcode = "OX108BB", radius = 2500, outputDir = ".")
}
\arguments{
\item{lcm_raster}{lcm raster. Defaults to files on W drive}

\item{dtm_raster}{lcm raster. Defaults to files on W drive}

\item{postcode}{Character, valid UK postcode to centre map on.}

\item{radius}{Numeric, radius of the map in meters}

\item{outputDir}{Character, path to output directory}
}
\value{
Path to the minecraft map
}
\description{
This wraps around other functions to create a minecraft world
from a postcode.
}
